Flattened build command
